Add vertiport / UAM edge data to b18EdgeData
Adds isVertiport, end_time_on_prev_edge, and getInterpolated
support so UAM trips can share the road-network sim loop.

+++ LivingCity/traffic/b18CUDA_trafficSimulator.h
@@ -18,7 +18,7 @@ extern void b18InitCUDA(bool fistInitialization, // crate buffers
                         std::vector<LC::B18TrafficPerson> &trafficPersonVec,
                         std::vector<uint> &indexPathVec,
                         std::vector<LC::B18EdgeData> &edgesData, std::vector<uchar> &laneMap,
-                        std::vector<uchar> &trafficLights,
+                        std::vector<float> &trafficLights,
                         std::vector<LC::B18IntersectionData> &intersections,
                         float startTimeH, float endTimeH,
                         std::vector<float>& accSpeedPerLinePerTimeInterval,
