+++ LivingCity/traffic/b18TrafficLaneMap.h
@@ -28,19 +28,19 @@ class B18TrafficLaneMap {
 
   void createLaneMap(const RoadGraph &inRoadGraph, std::vector<uchar> &laneMap,
       std::vector<B18EdgeData> &edgesData, std::vector<B18IntersectionData> &intersections,
-      std::vector<uchar> &trafficLights, std::map<uint,
+      std::vector<float> &trafficLights, std::map<uint,
       RoadGraph::roadGraphEdgeDesc_BI> &laneMapNumToEdgeDesc,
       std::map<RoadGraph::roadGraphEdgeDesc_BI, uint> &edgeDescToLaneMapNum);
 
   void createLaneMapSP(const std::shared_ptr<abm::Graph>& graph_, std::vector<uchar> &laneMap,
       std::vector<B18EdgeData> &edgesData, std::vector<B18IntersectionData> &intersections,
-      std::vector<uchar> &trafficLights, 
+      std::vector<float> &trafficLights, 
       std::map<uint, std::shared_ptr<abm::Graph::Edge>> &laneMapNumToEdgeDescSP,
       std::map<std::shared_ptr<abm::Graph::Edge>, uint> &edgeDescToLaneMapNumSP,
       std::vector<uint> &edgeIdToLaneMapNum);
 
   void resetIntersections(std::vector<B18IntersectionData> &intersections,
-                          std::vector<uchar> &trafficLights);
+                          std::vector<float> &trafficLights);
 
   
 };
